Integrate djb2 hash algorithm
Add the djb2 hash algorithm implementation for the hash function. This widely-used hashing algorithm is known for its simplicity and effectiveness in generating hash values for strings. The djb2 algorithm is a non-cryptographic hash function created by Daniel J. Bernstein and is implemented to provide efficient and uniform distribution of hash keys in the hash table data structure.